Factor src/imap/session.rs out of src/imap/client.rs
